T2D-66 Cleaning the code
