+++ src/main/java/uk/ac/ebi/ampt2d/accession/server/ApiError.java
@@ -21,26 +21,11 @@ import org.springframework.http.HttpStatus;
 
 public class ApiError {
 
-    private HttpStatus status;
-    private String message;
-    private String debugMessage;
-
-    private ApiError() {
-    }
-
-    ApiError(HttpStatus status) {
-        this.status = status;
-    }
-
-    ApiError(HttpStatus status, Throwable ex) {
-        this();
-        this.status = status;
-        this.message = "Unexpected error";
-        this.debugMessage = ex.getLocalizedMessage();
-    }
+    private final HttpStatus status;
+    private final String message;
+    private final String debugMessage;
 
     public ApiError(HttpStatus status, String message, Throwable ex) {
-        this();
         this.status = status;
         this.message = message;
         this.debugMessage = ex.getLocalizedMessage();
@@ -56,18 +41,6 @@ public class ApiError {
 
     public String getDebugMessage() {
         return debugMessage;
-    }
-
-    public void setStatus(HttpStatus status) {
-        this.status = status;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-
-    public void setDebugMessage(String debugMessage) {
-        this.debugMessage = debugMessage;
     }
 
     @Override
