+++ src/test/java/uk/ac/ebi/ampt2d/accession/file/UuidFileAccessioningServiceTest.java
@@ -20,7 +20,9 @@ package uk.ac.ebi.ampt2d.accession.file;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringRunner;
 
 import java.util.*;
@@ -28,7 +30,8 @@ import java.util.*;
 import static org.junit.Assert.*;
 
 @RunWith(SpringRunner.class)
-@SpringBootTest
+@DataJpaTest
+@ContextConfiguration(classes = FileConfiguration.class)
 public class UuidFileAccessioningServiceTest {
 
     @Autowired
